Convert BugReporter to an interface for testing purpose. Cannot use `@ContributesBinding(AppScope::class)`, so provide the implementation in AppModule.
