+++ features/rageshake/src/main/kotlin/io/element/android/features/rageshake/reporter/BugReporterListener.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2023 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.element.android.features.rageshake.reporter
+
+/**
+ * Bug report upload listener.
+ */
+interface BugReporterListener {
+    /**
+     * The bug report has been cancelled.
+     */
+    fun onUploadCancelled()
+
+    /**
+     * The bug report upload failed.
+     *
+     * @param reason the failure reason
+     */
+    fun onUploadFailed(reason: String?)
+
+    /**
+     * The upload progress (in percent).
+     *
+     * @param progress the upload progress
+     */
+    fun onProgress(progress: Int)
+
+    /**
+     * The bug report upload succeeded.
+     */
+    fun onUploadSucceed(reportUrl: String?)
+}
